Stop using refs to name fields of Records
This information is known in the Record itself.

This resolves an issue where Mems with a Record data type would run into
issues during FIRRTL emission or conversion when the refs of their
elements would be accessed to name the fields of the FIRRTL type.

This problematic behavior resulted from the following behaviors:
* Refs of Record elements are only set by _onModuleClose
* _onModuleClose is only called on Records added to the _ids of a Module
* Data are only added to _ids of Module on binding

The last point is new behavior in this PR